+++ R/color.spectro.R
@@ -3,10 +3,11 @@
 #' \code{color.spectro} highlights spectrogram regions specified by users
 #' @usage color.spectro(wave, wl = 512, wn = "hanning", ovlp = 70,
 #' dB = "max0", collevels = NULL, selec.col = "red2", col.clm = NULL, 
-#' base.col = "black", bg.col = "white", cexlab = 1, cexaxis = 1, tlab = "Time (s)", 
-#' flab = "Frequency (kHz)", title = NULL, axisX = TRUE, axisY = TRUE,
-#' flim = NULL, rm.zero = FALSE, X = NULL, fast.spec = FALSE, t.mar = NULL, f.mar = NULL,
-#' interactive = NULL, add = FALSE) 
+#' base.col = "black", bg.col = "white", strength = 1, 
+#' cexlab = 1, cexaxis = 1, tlab = "Time (s)", flab = "Frequency (kHz)", 
+#' title = NULL, axisX = TRUE, axisY = TRUE, flim = NULL, 
+#' rm.zero = FALSE, X = NULL, fast.spec = FALSE, t.mar = NULL, 
+#' f.mar = NULL, interactive = NULL, add = FALSE) 
 #' @param wave A 'wave' object produced by  \code{\link[tuneR]{readWave}} or similar functions.
 #' @param wl A numeric vector of length 1 specifying the window length of the spectrogram. Default 
 #'   is 512.
@@ -27,6 +28,7 @@
 #' Default is 'black'.
 #' @param bg.col Character vector of length 1 specifying the background color for both base
 #' and highlighted spectrograms. Default is 'white'.
+#' @param strength Numeric vector of length 1 controlling the strength of the highlighting color (actually how many times it is repeated in the internal color palette). Must be a positive integer. Default is 1.
 #' @param cexlab Numeric vector of length 1 specifying the relative size of axis 
 #'   labels. See \code{\link[seewave]{spectro}}. Default is 1.
 #' @param cexaxis Numeric vector of length 1 specifying the relative size of axis. See 
@@ -43,10 +45,10 @@
 #' @param  X Optional. Data frame containing columns for start and end time of signals ('start' and 'end') and low and high frequency ('bottom.freq' and 'top.freq'). 
 #' @param fast.spec Logical. If \code{TRUE} then image function is used internally to create spectrograms, which substantially 
 #' increases performance (much faster), although some options become unavailable, as collevels, and sc (amplitude scale).
-#' This option is indicated for signals with high background noise levels. Palette colors \code{\link[monitoR]{gray.1}}, \code{\link[monitoR]{gray.2}}, 
-#' \code{\link[monitoR]{gray.3}}, \code{\link[monitoR]{topo.1}} and \code{\link[monitoR]{rainbow.1}} (which should be imported from the package monitoR) seem
-#' to work better with 'fast' spectrograms. Palette colors \code{\link[monitoR]{gray.1}}, \code{\link[monitoR]{gray.2}}, 
-#' \code{\link[monitoR]{gray.3}} offer 
+#' This option is indicated for signals with high background noise levels. Palette colors \code{\link[monitoR:specCols]{gray.1}}, \code{\link[monitoR:specCols]{gray.2}}, 
+#' \code{\link[monitoR:specCols]{gray.3}}, \code{\link[monitoR:specCols]{topo.1}} and \code{\link[monitoR:specCols]{rainbow.1}} (which should be imported from the package monitoR) seem
+#' to work better with 'fast' spectrograms. Palette colors \code{\link[monitoR:specCols]{gray.1}}, \code{\link[monitoR:specCols]{gray.2}}, 
+#' \code{\link[monitoR:specCols]{gray.3}} offer 
 #' decreasing darkness levels. 
 #' @param t.mar Numeric vector of length 1. Specifies the margins adjacent to the start and end points to be added when highlighting selection. Default is \code{NULL}.
 #' @param f.mar Numeric vector of length 1. Specifies the margins adjacent to the low and high frequencies to be added when highlighting selection. Default is \code{NULL}.
@@ -79,22 +81,22 @@
 #'  st$colors <- c("red2", "blue", "green")
 #'  
 #'  # highlight selections
-#'  color.spectro(wave = sgnl, wl = 300, ovlp = 90, flim = c(1, 8.6), collevels = seq(-90, 0, 5), 
+#'  color.spectro(wave = sgnl, wl = 300, ovlp = 90, flim = c(1, 8.6), collevels = seq(-40, 0, 5), 
 #'  dB = "B", X = st, col.clm = "colors", base.col = "skyblue",  t.mar = 0.07, f.mar = 0.1, 
 #'  interactive = NULL)
 #'  
 #'  # interactive (selected manually: you have to select them by clicking on the spectrogram)
-#'  color.spectro(wave = sgnl, wl = 300, ovlp = 90, flim = c(1, 8.6), collevels = seq(-90, 0, 5),
+#'  color.spectro(wave = sgnl, wl = 300, ovlp = 90, flim = c(1, 8.6), collevels = seq(-40, 0, 5),
 #'   dB = "B", col.clm = "colors", t.mar = 0.07, f.mar = 1, interactive = 2)
 #' }
 #' 
 #' @references {Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.}
-#' @author Marcelo Araya-Salas (\email{marceloa27@@gmail.com}) and Grace Smith Vidaurre
+#' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr}) and Grace Smith Vidaurre
 #last modification on jul-5-2016 (MAS)
 
 color.spectro <- function(wave, wl = 512, wn = "hanning", ovlp = 70,
                           dB = "max0", collevels = NULL, selec.col = "red2", col.clm = NULL, 
-                          base.col = "black", bg.col = "white", cexlab = 1, cexaxis = 1, tlab = "Time (s)", 
+                          base.col = "black", bg.col = "white", strength = 1, cexlab = 1, cexaxis = 1, tlab = "Time (s)", 
                           flab = "Frequency (kHz)", title = NULL, axisX = TRUE, axisY = TRUE, 
                           flim = NULL, rm.zero = FALSE, X = NULL, fast.spec = FALSE, t.mar = NULL, f.mar = NULL,
                           interactive = NULL, add = FALSE) 
@@ -154,13 +156,13 @@ color.spectro <- function(wave, wl = 512, wn = "hanning", ovlp = 70,
     collevels <- seq(-40, 0, 5)
   
   # set background spectro color
-  basepal <- colorRampPalette(c(rep(bg.col, 2), base.col), alpha = TRUE) 
+  basepal <- colorRampPalette(c(bg.col, base.col)) 
   
   #adjust flim if lower than higher top.freq
-  if (!is.null(flim) & !is.null(sel.tab)) {
-    if (flim[1] > min(sel.tab$bottom.freq)) flim[1] <- min(sel.tab$bottom.freq)  
-    if (flim[2] < max(sel.tab$top.freq)) flim[2] <- max(sel.tab$top.freq)  
-  }
+  # if (!is.null(flim) & !is.null(sel.tab)) {
+  #   if (flim[1] > min(sel.tab$bottom.freq)) flim[1] <- min(sel.tab$bottom.freq)  
+  #   if (flim[2] < max(sel.tab$top.freq)) flim[2] <- max(sel.tab$top.freq)  
+  # }
   
   # add frequency margins
   if (!is.null(f.mar))
@@ -219,7 +221,10 @@ color.spectro <- function(wave, wl = 512, wn = "hanning", ovlp = 70,
     if (!is.null(dB)) 
       collevels <- seq(maxz - 30, maxz, by = 1) else   collevels <- seq(0, maxz, length = 30)
   }
-  
+  # opar <- par
+  # par(bg = bg.col)
+  # on.exit(par(opar), add = TRUE)
+  # 
   if (!fast.spec)
   {   # seewave spectro 
     #plot background spectro
@@ -227,8 +232,8 @@ color.spectro <- function(wave, wl = 512, wn = "hanning", ovlp = 70,
     filled_contour_color_wrblr_int(x = X, y = Y, z = Z, levels = collevels, 
                        nlevels = 20, plot.title = title(main = title, 
                                                         xlab = tlab, ylab = flab), color.palette = basepal, 
-                       axisX = axisX, axisY = axisY, col.lab = "black", 
-                       colaxis = "black")
+                       axisX = FALSE, axisY = axisY, col.lab = "black", 
+                       colaxis = "black", bg.col = bg.col)
     
     if (!is.null(interactive))
     {
@@ -247,9 +252,9 @@ color.spectro <- function(wave, wl = 512, wn = "hanning", ovlp = 70,
     #plot colored signals  
     if (!is.null(sel.tab))    
       out <- lapply(1:nrow(sel.tab), function(i)
-        filled_contour_color_wrblr_int(x = X[X > sel.tab$start[i] & X < sel.tab$end[i]], y = Y[Y > sel.tab$bottom.freq[i] & Y < sel.tab$top.freq[i]], z = Z[X > sel.tab$start[i] & X < sel.tab$end[i], Y > sel.tab$bottom.freq[i] & Y < sel.tab$top.freq[i]], nlevels = 20, plot.title = FALSE, color.palette = colorRampPalette(c(rep(bg.col, 2), colors[i]), alpha = TRUE), levels = collevels,
+        filled_contour_color_wrblr_int(x = X[X > sel.tab$start[i] & X < sel.tab$end[i]], y = Y[Y > sel.tab$bottom.freq[i] & Y < sel.tab$top.freq[i]], z = Z[X > sel.tab$start[i] & X < sel.tab$end[i], Y > sel.tab$bottom.freq[i] & Y < sel.tab$top.freq[i]], nlevels = 20, plot.title = FALSE, color.palette = colorRampPalette(c(rep(bg.col, 1), rep(colors[i], strength)), alpha = TRUE), levels = collevels,
                            axisX = FALSE, axisY = FALSE, col.lab = "black", 
-                           colaxis = "black", add = TRUE)   
+                           colaxis = "black", add = TRUE, bg.col = bg.col)   
       )
   } else {  #fast spectro image
     #plot background spectro
@@ -274,7 +279,7 @@ color.spectro <- function(wave, wl = 512, wn = "hanning", ovlp = 70,
   #plot colored signals
   if (!is.null(sel.tab))
         out <- lapply(1:nrow(sel.tab), function(i)
-                                       image(x = X[X > sel.tab$start[i] & X < sel.tab$end[i]], y = Y[Y > sel.tab$bottom.freq[i] & Y < sel.tab$top.freq[i]], z = Z[X > sel.tab$start[i] & X < sel.tab$end[i], Y > sel.tab$bottom.freq[i] & Y < sel.tab$top.freq[i]], col = colorRampPalette(c( rep(bg.col, 2), colors[i]), alpha = TRUE)(30), xlab = tlab, ylab = flab, axes = FALSE, xlim = range(X), add = TRUE)
+                                       image(x = X[X > sel.tab$start[i] & X < sel.tab$end[i]], y = Y[Y > sel.tab$bottom.freq[i] & Y < sel.tab$top.freq[i]], z = Z[X > sel.tab$start[i] & X < sel.tab$end[i], Y > sel.tab$bottom.freq[i] & Y < sel.tab$top.freq[i]], col = colorRampPalette(c( rep(bg.col, 2), rep(colors[i], strength)), alpha = TRUE)(30), xlab = tlab, ylab = flab, axes = FALSE, xlim = range(X), add = TRUE)
                                      )