+++ R/find_peaks.R
@@ -2,7 +2,7 @@
 #' 
 #' \code{find_peaks} find peaks in cross-correlation scores from \code{\link{xcorr}} 
 #' @usage find_peaks(xc.output, parallel = 1, cutoff = 0.4, path = NULL, pb = TRUE, 
-#' max.peak = FALSE)
+#' max.peak = FALSE, output = "data.frame")
 #' @param xc.output output of \code{\link{xcorr}} after setting \code{output = "list"}.
 #' @param parallel Numeric. Controls whether parallel computing is applied.
 #' It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).
@@ -11,6 +11,7 @@
 #' If \code{NULL} (default) then the current working directory is used.
 #' @param pb Logical argument to control progress bar. Default is \code{TRUE}.
 #' @param max.peak Logical argument to control whether only the peak with the highest correlation value is returned (if TRUE; cutoff will be ignored). Default is \code{FALSE}.
+#' @param output Character vector of length 1 to determine if only the detected peaks are returned ('cormat') or a list ('list') containing 1) the peaks  and 2) a data frame with correlation values at each sliding step for each comparison. The list, which is also of class 'peaks.output', can be used to graphically explore detections using \code{\link{lspec}}.
 #' @return The function returns a data frame with time and correlation score for the  
 #' detected peaks.
 #' @export
@@ -33,7 +34,7 @@
 #' pks <- find_peaks(xc.output = xc.output, path = tempdir())
 #' }
 #' @seealso \code{\link{autodetec}}, \code{\link[monitoR]{findPeaks}}
-#' @author Marcelo Araya-Salas \email{marceloa27@@gmail.com})
+#' @author Marcelo Araya-Salas \email{marcelo.araya@@ucr.ac.cr})
 #' 
 #' @references {
 #' Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.
@@ -41,7 +42,7 @@
 #' H. Khanna, S.L.L. Gaunt & D.A. McCallum (1997). Digital spectrographic cross-correlation: tests of sensitivity. Bioacoustics 7(3): 209-234
 #' }
 # last modification on jan-03-2020 (MAS)
-find_peaks <- function(xc.output, parallel = 1, cutoff = 0.4, path = NULL, pb = TRUE, max.peak = FALSE) 
+find_peaks <- function(xc.output, parallel = 1, cutoff = 0.4, path = NULL, pb = TRUE, max.peak = FALSE, output = "data.frame") 
 {
   # set pb options 
   on.exit(pbapply::pboptions(type = .Options$pboptions$type), add = TRUE)
@@ -76,8 +77,15 @@ find_peaks <- function(xc.output, parallel = 1, cutoff = 0.4, path = NULL, pb = 
       stop("'path' provided does not exist") else
         path <- normalizePath(path)
   
+  # set pb options 
+  pbapply::pboptions(type = ifelse(pb, "timer", "none"))
+  
+  # set clusters for windows OS and no soz
+  if (Sys.info()[1] == "Windows" & parallel > 1)
+    cl <- parallel::makePSOCKcluster(getOption("cl.cores", parallel)) else cl <- parallel
+  
   # loop over scores of each dyad
-  pks <- pbapply::pblapply(unique(xc.output$scores$dyad), FUN = function(i) {
+  pks <- pbapply::pblapply(unique(xc.output$scores$dyad), cl = cl, FUN = function(i) {
     
     # extract data for a dyad
     dat <- xc.output$scores[xc.output$scores$dyad == i, ]
@@ -142,7 +150,16 @@ find_peaks <- function(xc.output, parallel = 1, cutoff = 0.4, path = NULL, pb = 
   # sort columns in a intuitive order
   peaks <- sort_colms(peaks)
   
-  return(peaks)
+  # output results
+  if (output == "data.frame") return(peaks) else{
+    
+    output_list <- list(selection.table = peaks, scores = xc.output$scores,  cutoff = cutoff)
+    
+    class(output_list) <- c("list", "find_peaks.output")
+    
+    return(output_list)
+  }
+  
   } else {
   
   # no detections